Safari-609.1.11.1/LayoutTests/http/tests/frame-throttling/timer-throttle-in-cross-origin-subframe-expected.txt

Tests that DOM Timers are throttled in subframes that are cross-origin, and not in same-origin frames

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Received message: timerReachedMaxNestingLevel
Received message: timerReachedMaxNestingLevel
Checking that DOM Timers are throttled in cross origin frame
Received message: throttled[cross]: true
Received message: throttled[same]: false
PASS throttledState['cross'] is "true"
PASS throttledState['same'] is "false"
Interacted with cross-origin frame
Interacted with same-origin frame
Received message: throttled[cross]: false
Received message: throttled[same]: false
PASS throttledState['cross'] is "false"
PASS throttledState['same'] is "false"
PASS successfullyParsed is true

TEST COMPLETE
 
